# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Citas Médicas",
        "description": "Documentación de la API",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/addUser": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Crear usuario",
                "operationId": "7fcdd8801a80ed7884c4ccd69b781dea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthRegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login y obtención de token JWT",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthRegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation",
                    "rol"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "jx"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "julian@dd.d"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "12345"
                    },
                    "rol": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            },
            "AuthRegisterResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Usuario agregado correctamente"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "name": {
                                "type": "string",
                                "example": "Juan Pérez"
                            },
                            "email": {
                                "type": "string",
                                "example": "juan@example.com"
                            },
                            "rol": {
                                "type": "string",
                                "example": "admin"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AuthLoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "secret123"
                    }
                },
                "type": "object"
            },
            "AuthLoginResponse": {
                "properties": {
                    "Success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Bienvenido"
                    },
                    "Token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}